Update files (1.1.1.1) - New class: SliderConfigDouble, SliderDouble, SliderDoubleBuilder - Fix: SliderConfigInteger, SliderPercentBuilder, SliderIntegerBuilder

+++ Common/src/main/java/ru/kelcuprum/alinlib/gui/components/builder/slider/SliderPercentBuilder.java
@@ -74,7 +74,7 @@ public class SliderPercentBuilder {
         this.height = height;
         return this;
     }
-    public SliderPercentBuilder setDefaultValue(int defaultValue){
+    public SliderPercentBuilder setDefaultValue(double defaultValue){
         this.defaultValue = defaultValue;
         return this;
     }
